03_Data_ETL_Pipeline/project1_etl_api_to_db/requirements.txt
psycopg2-binary==2.9.5
requests==2.28.1
pandas==1.5.3
python-dotenv==0.21.0